+++ src/main/java/racingcar/view/RacingGameOutputView.java
@@ -1,7 +1,7 @@
 package racingcar.view;
 
-import racingcar.Car;
-import racingcar.Cars;
+import racingcar.domain.Car;
+import racingcar.domain.Cars;
 
 import java.util.List;
 import java.util.stream.Collectors;
