+++ src/test/java/racingcar/MovableStrategyTest.java
@@ -3,6 +3,7 @@ package racingcar;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.RepeatedTest;
 import org.junit.jupiter.api.Test;
+import racingcar.domain.MovableStrategy;
 
 public class MovableStrategyTest {
